Crate gstreamer_rtsp_server

Modules

prelude

Structs

Continue
Continue calling the closure in the future iterations or drop it.

Error
A generic error capable of representing various error domains (types).

RTSPAddress
An address

RTSPAddressFlags

RTSPAddressPool
An address pool, all member are private

RTSPAuth
The authentication structure.

RTSPClient
The client object represents the connection and its state with a client.

RTSPContext
Information passed around containing the context of a request.

RTSPMedia
A class that contains the GStreamer element along with a list of RTSPStream objects that can produce data.

RTSPMediaFactory
The definition and logic for constructing the pipeline for a media. The media can contain multiple streams like audio and video.

RTSPMediaFactoryURI
A media factory that creates a pipeline to play and uri.

RTSPMountPoints
Creates a RTSPMediaFactory object for a given url.

RTSPServer
This object listens on a port, creates and manages the clients connected to it.

RTSPSession
Session information kept by the server for a specific client. One client session, identified with a session id, can handle multiple medias identified with the url of a media.

RTSPSessionMedia
State of a client session regarding a specific media identified by path.

RTSPSessionPool
An object that keeps track of the active sessions. This object is usually attached to a RTSPServer object to manage the sessions in that server.

RTSPStream
The definition of a media stream.

RTSPStreamTransport
A Transport description for a stream

RTSPThreadPool
The thread pool structure.

RTSPTokenRef

RTSPTransportMode

RTSP_ADDRESS_POOL_ANY_IPV4

RTSP_ADDRESS_POOL_ANY_IPV6

RTSP_AUTH_CHECK_CONNECT

RTSP_AUTH_CHECK_MEDIA_FACTORY_ACCESS

RTSP_AUTH_CHECK_MEDIA_FACTORY_CONSTRUCT

RTSP_AUTH_CHECK_TRANSPORT_CLIENT_SETTINGS

RTSP_AUTH_CHECK_URL

RTSP_PERM_MEDIA_FACTORY_ACCESS

RTSP_PERM_MEDIA_FACTORY_CONSTRUCT

RTSP_TOKEN_MEDIA_FACTORY_ROLE

RTSP_TOKEN_TRANSPORT_CLIENT_SETTINGS

TypedValue
A statically typed Value.

Value
A generic value capable of carrying various types.

Enums

RTSPAddressPoolResult
Result codes from RTSP address pool functions.

RTSPMediaStatus
The state of the media pipeline.

RTSPPublishClockMode
Whether the clock and possibly RTP/clock offset should be published according to RFC7273.

RTSPSuspendMode
The suspend mode of the media pipeline. A media pipeline is suspended right after creating the SDP and when the client performs a PAUSED request.

RTSPThreadType
Different thread types

Type
A GLib or GLib-based library type

Traits

Cast
Upcasting and downcasting support.

GstRcRTSPTokenExt

IsA
Declares the "is a" relationship.

RTSPAddressPoolExt
Trait containing all RTSPAddressPool methods.

RTSPAddressPoolExtManual

RTSPAuthExt
Trait containing all RTSPAuth methods.

RTSPAuthExtManual

RTSPClientExt
Trait containing all RTSPClient methods.

RTSPClientExtManual

RTSPMediaExt
Trait containing all RTSPMedia methods.

RTSPMediaFactoryExt
Trait containing all RTSPMediaFactory methods.

RTSPMediaFactoryExtManual

RTSPMediaFactoryURIExt
Trait containing all RTSPMediaFactoryURI methods.

RTSPMountPointsExt
Trait containing all RTSPMountPoints methods.

RTSPServerExt
Trait containing all RTSPServer methods.

RTSPServerExtManual

RTSPSessionExt
Trait containing all RTSPSession methods.

RTSPSessionMediaExt
Trait containing all RTSPSessionMedia methods.

RTSPSessionPoolExt
Trait containing all RTSPSessionPool methods.

RTSPSessionPoolExtManual

RTSPStreamExt
Trait containing all RTSPStream methods.

RTSPStreamTransportExt
Trait containing all RTSPStreamTransport methods.

RTSPThreadPoolExt
Trait containing all RTSPThreadPool methods.

StaticType
Types that are supported by GLib dynamic typing.

ToValue
Converts to Value.

Type Definitions

RTSPToken