Trait gstreamer_rtsp_server::prelude::PresetExt

pub trait PresetExt {
    fn delete_preset(&self, name: &str) -> Result<(), BoolError>;

    fn get_meta(&self, name: &str, tag: &str) -> Option<String>;

    fn get_preset_names(&self) -> Vec<String>;

    fn get_property_names(&self) -> Vec<String>;

    fn is_editable(&self) -> bool;

    fn load_preset(&self, name: &str) -> Result<(), BoolError>;

    fn rename_preset(
        &self, 
        old_name: &str, 
        new_name: &str
    ) -> Result<(), BoolError>;

    fn save_preset(&self, name: &str) -> Result<(), BoolError>;

    fn set_meta<'a, P>(
        &self, 
        name: &str, 
        tag: &str, 
        value: P
    ) -> Result<(), BoolError>
    where
        P: Into<Option<&'a str>>;
}

Trait containing all Preset methods.

Implementors

Preset

Required Methods

fn delete_preset(&self, name: &str) -> Result<(), BoolError>

Delete the given preset.

name

preset name to remove

Returns

true for success, false if e.g. there is no preset with that name

fn get_meta(&self, name: &str, tag: &str) -> Option<String>

Gets the value for an existing meta data tag. Meta data tag names can be something like e.g. "comment". Returned values need to be released when done.

name

preset name

tag

meta data item name

value

value

Returns

true for success, false if e.g. there is no preset with that name or no value for the given tag

Important traits for Vec<u8>
impl Write for Vec<u8>

fn get_preset_names(&self) -> Vec<String>

Get a copy of preset names as a None terminated string array.

Returns

list with names, use g_strfreev after usage.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn get_property_names(&self) -> Vec<String>

Get a the names of the GObject properties that can be used for presets.

Returns

an array of property names which should be freed with g_strfreev after use.

fn is_editable(&self) -> bool

Check if one can add new presets, change existing ones and remove presets.

Returns

true if presets are editable or false if they are static

fn load_preset(&self, name: &str) -> Result<(), BoolError>

Load the given preset.

name

preset name to load

Returns

true for success, false if e.g. there is no preset with that name

fn rename_preset(&self, old_name: &str, new_name: &str) -> Result<(), BoolError>

Renames a preset. If there is already a preset by the new_name it will be overwritten.

old_name

current preset name

new_name

new preset name

Returns

true for success, false if e.g. there is no preset with old_name

fn save_preset(&self, name: &str) -> Result<(), BoolError>

Save the current object settings as a preset under the given name. If there is already a preset by this name it will be overwritten.

name

preset name to save

Returns

true for success, false

fn set_meta<'a, P>(
&self,
name: &str,
tag: &str,
value: P
) -> Result<(), BoolError> where
P: Into<Option<&'a str>>,

Sets a new value for an existing meta data item or adds a new item. Meta data tag names can be something like e.g. "comment". Supplying None for the value will unset an existing value.

name

preset name

tag

meta data item name

value

new value

Returns

true for success, false if e.g. there is no preset with that name