Trait gstreamer_rtsp_server::prelude::RTSPSessionMediaExt

pub trait RTSPSessionMediaExt {
    fn get_base_time(&self) -> ClockTime;

    fn get_media(&self) -> Option<RTSPMedia>;

    fn get_rtpinfo(&self) -> Option<String>;

    fn get_transport(&self, idx: u32) -> Option<RTSPStreamTransport>;

    fn matches(&self, path: &str) -> Option<i32>;

    fn set_state(&self, state: State) -> Result<(), BoolError>;
}

Trait containing all RTSPSessionMedia methods.

Implementors

RTSPSessionMedia

Required Methods

fn get_base_time(&self) -> ClockTime

Get the base_time of the RTSPMedia in self

Returns

the base_time of the media.

fn get_media(&self) -> Option<RTSPMedia>

Get the RTSPMedia that was used when constructing self

Returns

the RTSPMedia of self. Remains valid as long as self is valid.

fn get_rtpinfo(&self) -> Option<String>

Retrieve the RTP-Info header string for all streams in self with configured transports.

Returns

The RTP-Info as a string or None when no RTP-Info could be generated, g_free after usage.

fn get_transport(&self, idx: u32) -> Option<RTSPStreamTransport>

Get a previously created RTSPStreamTransport for the stream at idx.

idx

the stream index

Returns

a RTSPStreamTransport that is valid until the session of self is unreffed.

fn matches(&self, path: &str) -> Option<i32>

Check if the path of self matches path. It path matches, the amount of matched characters is returned in matched.

path

a path

matched

the amount of matched characters of path

Returns

true when path matches the path of self.

fn set_state(&self, state: State) -> Result<(), BoolError>

Tell the media object self to change to state.

state

the new state

Returns

true on success.