Trait gstreamer_rtsp_server::prelude::TagSetterExt

pub trait TagSetterExt {
    fn get_tag_list(&self) -> Option<GstRc<TagListRef>>;

    fn get_tag_merge_mode(&self) -> TagMergeMode;

    fn merge_tags(&self, list: &GstRc<TagListRef>, mode: TagMergeMode);

    fn reset_tags(&self);

    fn set_tag_merge_mode(&self, mode: TagMergeMode);
}

Trait containing all TagSetter methods.

Implementors

TagSetter

Required Methods

fn get_tag_list(&self) -> Option<GstRc<TagListRef>>

Returns the current list of tags the setter uses. The list should not be modified or freed.

This function is not thread-safe.

Returns

a current snapshot of the taglist used in the setter or None if none is used.

fn get_tag_merge_mode(&self) -> TagMergeMode

Queries the mode by which tags inside the setter are overwritten by tags from events

Returns

the merge mode used inside the element.

fn merge_tags(&self, list: &GstRc<TagListRef>, mode: TagMergeMode)

Merges the given list into the setter's list using the given mode.

list

a tag list to merge from

mode

the mode to merge with

fn reset_tags(&self)

Reset the internal taglist. Elements should call this from within the state-change handler.

fn set_tag_merge_mode(&self, mode: TagMergeMode)

Sets the given merge mode that is used for adding tags from events to tags specified by this interface. The default is TagMergeMode::Keep, which keeps the tags set with this interface and discards tags from events.

mode

The mode with which tags are added