Trait gstreamer_rtsp_server::RTSPClientExt

pub trait RTSPClientExt {
    fn close(&self);

    fn get_auth(&self) -> Option<RTSPAuth>;

    fn get_mount_points(&self) -> Option<RTSPMountPoints>;

    fn get_session_pool(&self) -> Option<RTSPSessionPool>;

    fn get_thread_pool(&self) -> Option<RTSPThreadPool>;

    fn set_auth<'a, P: Into<Option<&'a RTSPAuth>>>(&self, auth: P);

    fn set_mount_points<'a, P: Into<Option<&'a RTSPMountPoints>>>(
        &self, 
        mounts: P
    );

    fn set_session_pool<'a, P: Into<Option<&'a RTSPSessionPool>>>(
        &self, 
        pool: P
    );

    fn set_thread_pool<'a, P: Into<Option<&'a RTSPThreadPool>>>(&self, pool: P);

    fn get_property_drop_backlog(&self) -> bool;

    fn set_property_drop_backlog(&self, drop_backlog: bool);

    fn connect_announce_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_closed<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_describe_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_get_parameter_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_handle_response<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_new_session<F: Fn(&Self, &RTSPSession) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_options_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pause_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_play_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pre_announce_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pre_describe_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pre_get_parameter_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pre_options_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pre_pause_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pre_play_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pre_record_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pre_set_parameter_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pre_setup_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_pre_teardown_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_record_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_set_parameter_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_setup_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_teardown_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_drop_backlog_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_mount_points_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_session_pool_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all RTSPClient methods.

Implementors

RTSPClient

Required Methods

fn close(&self)

Close the connection of self and remove all media it was managing.

fn get_auth(&self) -> Option<RTSPAuth>

Get the RTSPAuth used as the authentication manager of self.

Returns

the RTSPAuth of self. gobject::ObjectExt::unref after usage.

fn get_mount_points(&self) -> Option<RTSPMountPoints>

Get the RTSPMountPoints object that self uses to manage its sessions.

Returns

a RTSPMountPoints, unref after usage.

fn get_session_pool(&self) -> Option<RTSPSessionPool>

Get the RTSPSessionPool object that self uses to manage its sessions.

Returns

a RTSPSessionPool, unref after usage.

fn get_thread_pool(&self) -> Option<RTSPThreadPool>

Get the RTSPThreadPool used as the thread pool of self.

Returns

the RTSPThreadPool of self. gobject::ObjectExt::unref after usage.

fn set_auth<'a, P: Into<Option<&'a RTSPAuth>>>(&self, auth: P)

configure auth to be used as the authentication manager of self.

auth

a RTSPAuth

fn set_mount_points<'a, P: Into<Option<&'a RTSPMountPoints>>>(&self, mounts: P)

Set mounts as the mount points for self which it will use to map urls to media streams. These mount points are usually inherited from the server that created the client but can be overriden later.

mounts

a RTSPMountPoints

fn set_session_pool<'a, P: Into<Option<&'a RTSPSessionPool>>>(&self, pool: P)

Set pool as the sessionpool for self which it will use to find or allocate sessions. the sessionpool is usually inherited from the server that created the client but can be overridden later.

pool

a RTSPSessionPool

fn set_thread_pool<'a, P: Into<Option<&'a RTSPThreadPool>>>(&self, pool: P)

configure pool to be used as the thread pool of self.

pool

a RTSPThreadPool

fn get_property_drop_backlog(&self) -> bool

fn set_property_drop_backlog(&self, drop_backlog: bool)

fn connect_announce_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_closed<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_describe_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_get_parameter_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_handle_response<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_new_session<F: Fn(&Self, &RTSPSession) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_options_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_pause_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_play_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_pre_announce_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Feature: v1_12

ctx

a RTSPContext

Returns

a gst_rtsp::RTSPStatusCode, GST_RTSP_STS_OK in case of success, otherwise an appropriate return code

fn connect_pre_describe_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Feature: v1_12

ctx

a RTSPContext

Returns

a gst_rtsp::RTSPStatusCode, GST_RTSP_STS_OK in case of success, otherwise an appropriate return code

fn connect_pre_get_parameter_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Feature: v1_12

ctx

a RTSPContext

Returns

a gst_rtsp::RTSPStatusCode, GST_RTSP_STS_OK in case of success, otherwise an appropriate return code

fn connect_pre_options_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Feature: v1_12

ctx

a RTSPContext

Returns

a gst_rtsp::RTSPStatusCode, GST_RTSP_STS_OK in case of success, otherwise an appropriate return code

fn connect_pre_pause_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Feature: v1_12

ctx

a RTSPContext

Returns

a gst_rtsp::RTSPStatusCode, GST_RTSP_STS_OK in case of success, otherwise an appropriate return code

fn connect_pre_play_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Feature: v1_12

ctx

a RTSPContext

Returns

a gst_rtsp::RTSPStatusCode, GST_RTSP_STS_OK in case of success, otherwise an appropriate return code

fn connect_pre_record_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Feature: v1_12

ctx

a RTSPContext

Returns

a gst_rtsp::RTSPStatusCode, GST_RTSP_STS_OK in case of success, otherwise an appropriate return code

fn connect_pre_set_parameter_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Feature: v1_12

ctx

a RTSPContext

Returns

a gst_rtsp::RTSPStatusCode, GST_RTSP_STS_OK in case of success, otherwise an appropriate return code

fn connect_pre_setup_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Feature: v1_12

ctx

a RTSPContext

Returns

a gst_rtsp::RTSPStatusCode, GST_RTSP_STS_OK in case of success, otherwise an appropriate return code

fn connect_pre_teardown_request<F: Fn(&Self, &RTSPContext) -> RTSPStatusCode + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

Feature: v1_12

ctx

a RTSPContext

Returns

a gst_rtsp::RTSPStatusCode, GST_RTSP_STS_OK in case of success, otherwise an appropriate return code

fn connect_record_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_set_parameter_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_setup_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_teardown_request<F: Fn(&Self, &RTSPContext) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

ctx

a RTSPContext

fn connect_property_drop_backlog_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_mount_points_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_session_pool_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId