Trait gstreamer_rtsp_server::RTSPMediaFactoryExt

pub trait RTSPMediaFactoryExt {
    fn construct(&self, url: &RTSPUrl) -> Option<RTSPMedia>;

    fn create_element(&self, url: &RTSPUrl) -> Option<Element>;

    fn get_address_pool(&self) -> Option<RTSPAddressPool>;

    fn get_buffer_size(&self) -> u32;

    fn get_clock(&self) -> Option<Clock>;

    fn get_latency(&self) -> u32;

    fn get_launch(&self) -> Option<String>;

    fn get_media_gtype(&self) -> Type;

    fn get_multicast_iface(&self) -> Option<String>;

    fn get_profiles(&self) -> RTSPProfile;

    fn get_protocols(&self) -> RTSPLowerTrans;

    fn get_publish_clock_mode(&self) -> RTSPPublishClockMode;

    fn get_retransmission_time(&self) -> ClockTime;

    fn get_suspend_mode(&self) -> RTSPSuspendMode;

    fn get_transport_mode(&self) -> RTSPTransportMode;

    fn is_eos_shutdown(&self) -> bool;

    fn is_shared(&self) -> bool;

    fn is_stop_on_disonnect(&self) -> bool;

    fn set_address_pool<'a, P: Into<Option<&'a RTSPAddressPool>>>(
        &self, 
        pool: P
    );

    fn set_buffer_size(&self, size: u32);

    fn set_clock<'a, P: IsA<Clock> + 'a, Q: Into<Option<&'a P>>>(
        &self, 
        clock: Q
    );

    fn set_eos_shutdown(&self, eos_shutdown: bool);

    fn set_latency(&self, latency: u32);

    fn set_launch(&self, launch: &str);

    fn set_media_gtype(&self, media_gtype: Type);

    fn set_multicast_iface<'a, P: Into<Option<&'a str>>>(
        &self, 
        multicast_iface: P
    );

    fn set_profiles(&self, profiles: RTSPProfile);

    fn set_protocols(&self, protocols: RTSPLowerTrans);

    fn set_publish_clock_mode(&self, mode: RTSPPublishClockMode);

    fn set_retransmission_time(&self, time: ClockTime);

    fn set_shared(&self, shared: bool);

    fn set_stop_on_disconnect(&self, stop_on_disconnect: bool);

    fn set_suspend_mode(&self, mode: RTSPSuspendMode);

    fn set_transport_mode(&self, mode: RTSPTransportMode);

    fn get_property_eos_shutdown(&self) -> bool;

    fn get_property_shared(&self) -> bool;

    fn get_property_stop_on_disconnect(&self) -> bool;

    fn connect_media_configure<F: Fn(&Self, &RTSPMedia) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_media_constructed<F: Fn(&Self, &RTSPMedia) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_buffer_size_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_clock_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_eos_shutdown_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_latency_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_launch_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_profiles_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_protocols_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_shared_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_stop_on_disconnect_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_suspend_mode_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_transport_mode_notify<F: Fn(&Self) + Send + Sync + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all RTSPMediaFactory methods.

Implementors

RTSPMediaFactoryURI, RTSPMediaFactory

Required Methods

fn construct(&self, url: &RTSPUrl) -> Option<RTSPMedia>

Construct the media object and create its streams. Implementations should create the needed gstreamer elements and add them to the result object. No state changes should be performed on them yet.

One or more GstRTSPStream objects should be created from the result with gst_rtsp_media_create_stream ().

After the media is constructed, it can be configured and then prepared with gst_rtsp_media_prepare ().

url

the url used

Returns

a new RTSPMedia if the media could be prepared.

fn create_element(&self, url: &RTSPUrl) -> Option<Element>

Construct and return a gst::Element that is a gst::Bin containing the elements to use for streaming the media.

The bin should contain payloaders pay\%d for each stream. The default implementation of this function returns the bin created from the launch parameter.

url

the url used

Returns

a new gst::Element.

fn get_address_pool(&self) -> Option<RTSPAddressPool>

Get the RTSPAddressPool used as the address pool of self.

Returns

the RTSPAddressPool of self. gobject::ObjectExt::unref after usage.

fn get_buffer_size(&self) -> u32

Get the kernel UDP buffer size.

Returns

the kernel UDP buffer size.

fn get_clock(&self) -> Option<Clock>

Returns the clock that is going to be used by the pipelines of all medias created from this factory.

Returns

The GstClock

fn get_latency(&self) -> u32

Get the latency that is used for receiving media

Returns

latency in milliseconds

fn get_launch(&self) -> Option<String>

Get the gst_parse_launch pipeline description that will be used in the default prepare vmethod.

Returns

the configured launch description. g_free after usage.

fn get_media_gtype(&self) -> Type

Return the GType of the GstRTSPMedia subclass this factory will create.

fn get_multicast_iface(&self) -> Option<String>

Get the multicast interface used for self.

Returns

the multicast interface for self. g_free after usage.

fn get_profiles(&self) -> RTSPProfile

Get the allowed profiles of self.

Returns

a gst_rtsp::RTSPProfile

fn get_protocols(&self) -> RTSPLowerTrans

Get the allowed protocols of self.

Returns

a gst_rtsp::RTSPLowerTrans

fn get_publish_clock_mode(&self) -> RTSPPublishClockMode

Gets if and how the media clock should be published according to RFC7273.

Returns

The GstRTSPPublishClockMode

fn get_retransmission_time(&self) -> ClockTime

Get the time that is stored for retransmission purposes

Returns

a gst::ClockTime

fn get_suspend_mode(&self) -> RTSPSuspendMode

Get how media created from this factory will be suspended.

Returns

a RTSPSuspendMode.

fn get_transport_mode(&self) -> RTSPTransportMode

Get if media created from this factory can be used for PLAY or RECORD methods.

Returns

The transport mode.

fn is_eos_shutdown(&self) -> bool

Get if media created from this factory will have an EOS event sent to the pipeline before shutdown.

Returns

true if the media will receive EOS before shutdown.

fn is_shared(&self) -> bool

Get if media created from this factory can be shared between clients.

Returns

true if the media will be shared between clients.

fn is_stop_on_disonnect(&self) -> bool

fn set_address_pool<'a, P: Into<Option<&'a RTSPAddressPool>>>(&self, pool: P)

configure pool to be used as the address pool of self.

pool

a RTSPAddressPool

fn set_buffer_size(&self, size: u32)

Set the kernel UDP buffer size.

size

the new value

fn set_clock<'a, P: IsA<Clock> + 'a, Q: Into<Option<&'a P>>>(&self, clock: Q)

Configures a specific clock to be used by the pipelines of all medias created from this factory.

clock

the clock to be used by the media factory

fn set_eos_shutdown(&self, eos_shutdown: bool)

Configure if media created from this factory will have an EOS sent to the pipeline before shutdown.

eos_shutdown

the new value

fn set_latency(&self, latency: u32)

Configure the latency used for receiving media

latency

latency in milliseconds

fn set_launch(&self, launch: &str)

The gst_parse_launch line to use for constructing the pipeline in the default prepare vmethod.

The pipeline description should return a GstBin as the toplevel element which can be accomplished by enclosing the description with brackets '(' ')'.

The description should return a pipeline with payloaders named pay0, pay1, etc.. Each of the payloaders will result in a stream.

launch

the launch description

fn set_media_gtype(&self, media_gtype: Type)

Configure the GType of the GstRTSPMedia subclass to create (by default, overridden construct vmethods may of course do something different)

media_gtype

the GType of the class to create

fn set_multicast_iface<'a, P: Into<Option<&'a str>>>(&self, multicast_iface: P)

configure multicast_iface to be used for self.

multicast_iface

a multicast interface name

fn set_profiles(&self, profiles: RTSPProfile)

Configure the allowed profiles for self.

profiles

the new flags

fn set_protocols(&self, protocols: RTSPLowerTrans)

Configure the allowed lower transport for self.

protocols

the new flags

fn set_publish_clock_mode(&self, mode: RTSPPublishClockMode)

Sets if and how the media clock should be published according to RFC7273.

mode

the clock publish mode

fn set_retransmission_time(&self, time: ClockTime)

Configure the time to store for possible retransmission

time

a gst::ClockTime

fn set_shared(&self, shared: bool)

Configure if media created from this factory can be shared between clients.

shared

the new value

fn set_stop_on_disconnect(&self, stop_on_disconnect: bool)

Configure if media created from this factory should be stopped when a client disconnects without sending TEARDOWN.

stop_on_disconnect

the new value

fn set_suspend_mode(&self, mode: RTSPSuspendMode)

Configure how media created from this factory will be suspended.

mode

the new RTSPSuspendMode

fn set_transport_mode(&self, mode: RTSPTransportMode)

Configure if this factory creates media for PLAY or RECORD modes.

mode

the new value

fn get_property_eos_shutdown(&self) -> bool

fn get_property_shared(&self) -> bool

fn get_property_stop_on_disconnect(&self) -> bool

fn connect_media_configure<F: Fn(&Self, &RTSPMedia) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_media_constructed<F: Fn(&Self, &RTSPMedia) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_buffer_size_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_clock_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_eos_shutdown_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_latency_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_launch_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_profiles_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_protocols_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_shared_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_stop_on_disconnect_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_suspend_mode_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_transport_mode_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId