Trait gstreamer_rtsp_server::RTSPStreamTransportExt

pub trait RTSPStreamTransportExt {
    fn get_rtpinfo(&self, start_time: ClockTime) -> Option<String>;

    fn get_stream(&self) -> Option<RTSPStream>;

    fn get_url(&self) -> Option<RTSPUrl>;

    fn is_timed_out(&self) -> bool;

    fn keep_alive(&self);

    fn recv_data(&self, channel: u32, buffer: &Buffer) -> FlowReturn;

    fn send_rtcp(&self, buffer: &Buffer) -> Result<(), BoolError>;

    fn send_rtp(&self, buffer: &Buffer) -> Result<(), BoolError>;

    fn set_active(&self, active: bool) -> Result<(), BoolError>;

    fn set_timed_out(&self, timedout: bool);

    fn set_url<'a, P: Into<Option<&'a RTSPUrl>>>(&self, url: P);
}

Trait containing all RTSPStreamTransport methods.

Implementors

RTSPStreamTransport

Required Methods

fn get_rtpinfo(&self, start_time: ClockTime) -> Option<String>

Get the RTP-Info string for self and start_time.

start_time

a star time

Returns

the RTPInfo string for self and start_time or None when the RTP-Info could not be determined. g_free after usage.

fn get_stream(&self) -> Option<RTSPStream>

Get the RTSPStream used when constructing self.

Returns

the stream used when constructing self.

fn get_url(&self) -> Option<RTSPUrl>

Get the url configured in self.

Returns

the url configured in self. It remains valid for as long as self is valid.

fn is_timed_out(&self) -> bool

Check if self is timed out.

Returns

true if self timed out.

fn keep_alive(&self)

Signal the installed keep_alive callback for self.

fn recv_data(&self, channel: u32, buffer: &Buffer) -> FlowReturn

Receive buffer on channel self.

channel

a channel

buffer

a gst::Buffer

Returns

a gst::FlowReturn. Returns GST_FLOW_NOT_LINKED when channel is not configured in the transport of self.

fn send_rtcp(&self, buffer: &Buffer) -> Result<(), BoolError>

Send buffer to the installed RTCP callback for self.

buffer

a gst::Buffer

Returns

true on success

fn send_rtp(&self, buffer: &Buffer) -> Result<(), BoolError>

Send buffer to the installed RTP callback for self.

buffer

a gst::Buffer

Returns

true on success

fn set_active(&self, active: bool) -> Result<(), BoolError>

Activate or deactivate datatransfer configured in self.

active

new state of self

Returns

true when the state was changed.

fn set_timed_out(&self, timedout: bool)

Set the timed out state of self to timedout

timedout

timed out value

fn set_url<'a, P: Into<Option<&'a RTSPUrl>>>(&self, url: P)

Set url as the client url.

url

a client gst_rtsp::RTSPUrl